requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
decorator==4.0.10
docopt==0.6.2
funcsigs==1.0.2
ipython==5.1.0
ipython-genutils==0.1.0
matplotlib==1.5.2
mock==2.0.0
numpy==1.11.1
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
protobuf==3.0.0b2
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.8
pyserial==3.1.1
python-dateutil==2.5.3
pytz==2016.6.1
simplegeneric==0.8.1
six==1.10.0
tensorflow==0.10.0rc0
traitlets==4.2.2
wcwidth==0.1.7
